Changed: Cleanup README and Move `traits` module for format handlers into `handlers`.
