+++ projects/api/dxt-lossless-transform-file-formats-api/src/api.rs
@@ -1,10 +1,8 @@
 //! High-level convenience APIs for file format operations.
 
 use crate::bundle::TransformBundle;
-use crate::embed::formats::{EmbeddableBc1Details, EmbeddableTransformDetails};
-use crate::embed::{TransformFormat, TransformHeader};
 use crate::error::{FormatHandlerError, TransformError, TransformResult};
-use crate::traits::file_format_handler::FileFormatHandler;
+use crate::handlers::{FileFormatDetection, FileFormatHandler, FileFormatUntransformDetection};
 use core::fmt::Debug;
 use dxt_lossless_transform_api_common::estimate::SizeEstimationOperations;
 
@@ -102,113 +100,6 @@ pub fn untransform_slice<H: FileFormatHandler>(
     handler.untransform(input, output)
 }
 
-/// Dispatch untransform operation based on the transform header format.
-///
-/// This is a lower-level function that operates directly on texture data,
-/// assuming the file format headers have already been processed.
-///
-/// # Parameters
-///
-/// - `header`: The transform header containing format and settings
-/// - `input_texture_data`: Input slice containing the transformed texture data
-/// - `output_texture_data`: Output slice where the untransformed texture data will be written (must be at least the same size as input)
-///
-/// # Safety Requirements
-///
-/// Both input and output texture data must be properly sized for the format:
-/// - BC1: Must be multiple of 8 bytes
-/// - BC2/BC3: Must be multiple of 16 bytes  
-/// - BC7: Must be multiple of 16 bytes
-///
-/// Output buffer must be at least the same size as the input buffer.
-///
-/// # Example
-///
-/// See: `dxt-lossless-transform-dds` crate.
-pub fn dispatch_untransform(
-    header: TransformHeader,
-    input_texture_data: &[u8],
-    output_texture_data: &mut [u8],
-) -> TransformResult<()> {
-    if output_texture_data.len() < input_texture_data.len() {
-        return Err(TransformError::FormatHandler(
-            FormatHandlerError::OutputBufferTooSmall {
-                required: input_texture_data.len(),
-                actual: output_texture_data.len(),
-            },
-        ));
-    }
-
-    match header.format() {
-        Some(TransformFormat::Bc1) => {
-            let details = EmbeddableBc1Details::from_header(header)?;
-
-            // BC1 untransform using unsafe API with safe wrapper
-            if input_texture_data.len() % 8 != 0 {
-                return Err(TransformError::InvalidDataAlignment {
-                    size: input_texture_data.len(),
-                    required_divisor: 8,
-                });
-            }
-
-            unsafe {
-                dxt_lossless_transform_bc1::untransform_bc1_with_settings(
-                    input_texture_data.as_ptr(),
-                    output_texture_data.as_mut_ptr(),
-                    input_texture_data.len(),
-                    details.to_settings(),
-                );
-            }
-        }
-        _ => {
-            return Err(TransformError::UnknownTransformFormat);
-        }
-    }
-
-    Ok(())
-}
-
-/// Dispatch transform operation based on the detected format.
-///
-/// This is a lower-level function that operates directly on texture data,
-/// assuming the file format headers have already been processed.
-///
-/// # Parameters
-///
-/// - `format`: The detected texture format to transform
-/// - `input_texture_data`: Input slice containing the original texture data
-/// - `output_texture_data`: Output slice where the transformed texture data will be written (must be at least the same size as input)
-/// - `bundle`: Bundle containing transform builders for different BCx formats
-///
-/// # Returns
-///
-/// Returns a [`TransformHeader`] containing the transform details that should be embedded in the file.
-///
-/// # Safety Requirements
-///
-/// Both input and output texture data must be properly sized for the format:
-/// - BC1: Must be multiple of 8 bytes
-/// - BC2/BC3: Must be multiple of 16 bytes  
-/// - BC7: Must be multiple of 16 bytes
-///
-/// Output buffer must be at least the same size as the input buffer.
-///
-/// # Example
-///
-/// See: `dxt-lossless-transform-dds` crate.
-pub fn dispatch_transform<T>(
-    format: TransformFormat,
-    input_texture_data: &[u8],
-    output_texture_data: &mut [u8],
-    bundle: &TransformBundle<T>,
-) -> TransformResult<TransformHeader>
-where
-    T: SizeEstimationOperations,
-    T::Error: Debug,
-{
-    bundle.dispatch_transform(format, input_texture_data, output_texture_data)
-}
-
 /// Transform a slice using multiple handlers with automatic format detection.
 ///
 /// This function tries each handler in sequence until one accepts the file format,
@@ -216,7 +107,7 @@ where
 ///
 /// # Parameters
 ///
-/// - `handlers`: Iterator of file format handlers that implement [`crate::traits::FileFormatDetection`]
+/// - `handlers`: Iterator of file format handlers that implement [`FileFormatDetection`]
 /// - `input`: Input buffer containing the file data
 /// - `output`: Output buffer (must be at least the same size as input)
 /// - `bundle`: Bundle containing transform builders for different BCx formats
@@ -248,7 +139,7 @@ pub fn transform_slice_with_multiple_handlers<HandlerIterator, Handler, SizeEsti
 ) -> TransformResult<()>
 where
     HandlerIterator: IntoIterator<Item = Handler>,
-    Handler: crate::traits::FileFormatDetection,
+    Handler: FileFormatDetection,
     SizeEstimator: SizeEstimationOperations,
     SizeEstimator::Error: Debug,
 {
@@ -279,7 +170,7 @@ where
 ///
 /// # Parameters
 ///
-/// - `handlers`: Iterator of file format handlers that implement [`crate::traits::FileFormatUntransformDetection`]
+/// - `handlers`: Iterator of file format handlers that implement [`FileFormatUntransformDetection`]
 /// - `input`: Input buffer containing transformed data
 /// - `output`: Output buffer (must be at least the same size as input)
 ///
@@ -307,7 +198,7 @@ pub fn untransform_slice_with_multiple_handlers<HandlerIterator, Handler>(
 ) -> TransformResult<()>
 where
     HandlerIterator: IntoIterator<Item = Handler>,
-    Handler: crate::traits::FileFormatUntransformDetection,
+    Handler: FileFormatUntransformDetection,
 {
     if output.len() < input.len() {
         return Err(TransformError::FormatHandler(
@@ -363,19 +254,6 @@ mod tests {
         // Verify that the handler was called
         let calls = handler.get_calls();
         assert!(calls.untransform_called);
-    }
-
-    #[test]
-    fn test_dispatch_untransform_invalid_alignment() {
-        let header = create_test_bc1_header();
-        let input = vec![0u8; 15]; // Not multiple of 8
-        let mut output = vec![0u8; 15];
-
-        let result = dispatch_untransform(header, &input, &mut output);
-        assert!(matches!(
-            result,
-            Err(TransformError::InvalidDataAlignment { .. })
-        ));
     }
 
     #[test]
