+++ projects/extensions/dxt-lossless-transform-dds/src/lib.rs
@@ -8,7 +8,7 @@ extern crate std;
 pub mod test_prelude;
 
 pub mod dds;
-pub mod format_handler;
+pub mod handler;
 
 // Re-export the DDS handler for convenient access
-pub use format_handler::DdsHandler;
+pub use handler::DdsHandler;
